+++ src/us/kbase/workspace/FuncInfo.java
@@ -15,22 +15,8 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 /**
  * <p>Original spec-file type: FuncInfo</p>
  * <pre>
- * Information about a function
- *         func_string func_def - resolved func definition id.
- *         string description - the description of the function from spec file.
- *         string spec_def - reconstruction of function definition from spec file.
- *         string parsing_structure - json document describing parsing structure of function 
- *                 in spec file including types of arguments.
- *         list<spec_version> module_vers - versions of spec files containing
- *                 given func version.
- *         list<spec_version> released_module_vers - released versions of spec files 
- *                 containing given func version.
- *         list<func_string> func_vers - all versions of function with given type
- *                 name.
- *         list<func_string> released_func_vers - all released versions of function 
- *                 with given type name.
- *         list<type_string> used_type_defs - list of types (with versions) 
- *                 referred to from this function version.
+ * DEPRECATED
+ * @deprecated
  * </pre>
  * 
  */