+++ src/us/kbase/workspace/test/kbase/JSONRPCLayerTest.java
@@ -4408,7 +4408,9 @@ public class JSONRPCLayerTest extends JSONRPCLayerTester {
 		assertThat(cl.getTypeInfo("UnreleasedModule.AType").getTypeVers().size(), is(1));
 		assertThat(cl.getTypeInfo("UnreleasedModule.AType-0.1").getTypeVers().size(), is(1));
 		assertThat(cl.getJsonschema("UnreleasedModule.AType").length() > 0, is(true));
-		assertThat(cl.getFuncInfo("UnreleasedModule.aFunc").getFuncVers().size(), is(1));
+		@SuppressWarnings("deprecation")
+		final List<String> funcVers = cl.getFuncInfo("UnreleasedModule.aFunc").getFuncVers();
+		assertThat(funcVers.size(), is(1));
 		try {
 			cl.getTypeInfo("UnreleasedModule.AType-0.2");
 			fail();
@@ -4517,8 +4519,9 @@ public class JSONRPCLayerTest extends JSONRPCLayerTester {
 				is("TestModule.Feature-1.0"));
 
 		// make sure we can get func info
-		assertThat(CLIENT1.getFuncInfo("TestModule.getFeature").getFuncDef(),
-				is("TestModule.getFeature-1.0"));
+		@SuppressWarnings("deprecation")
+		final String funcDef = CLIENT1.getFuncInfo("TestModule.getFeature").getFuncDef();
+		assertThat(funcDef, is("TestModule.getFeature-1.0"));
 	}
 
 	@Test
@@ -4577,7 +4580,9 @@ public class JSONRPCLayerTest extends JSONRPCLayerTester {
 	@Test
 	public void testGetAllTypeAndFuncInfo() throws Exception {
 		assertThat(CLIENT1.getAllTypeInfo("RefSpec").size(), is(1));
-		assertThat(CLIENT_FOR_SRV2.getAllFuncInfo("UnreleasedModule").size(), is(1));
+		@SuppressWarnings("deprecation")
+		final int size = CLIENT_FOR_SRV2.getAllFuncInfo("UnreleasedModule").size();
+		assertThat(size, is(1));
 	}
 
 	@Test